backend/requirements.txt
psycopg2-binary
flask
numpy